# openapi.yaml
openapi: 3.0.4
info:
  title: Strava API v3
  description: FastAPI implementation of Strava API v3
  version: 3.0.0
paths:
...
  /athlete/activities:
    get:
      tags:
        - Athlete
      summary: Get Athlete Activities
      description: Returns the activities of the authenticated athlete.
      operationId: getAthleteActivities
      parameters:
        - name: before
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
            nullable: true
            description: Filter activities before this timestamp
            title: Before
          description: Filter activities before this timestamp
        - name: after
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
            nullable: true
            description: Filter activities after this timestamp
            title: After
          description: Filter activities after this timestamp
        - name: page
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
            nullable: true
            description: Page number
            default: 1
            title: Page
          description: Page number
        - name: per_page
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
            nullable: true
            description: Number of items per page
            default: 30
            title: Per Page
          description: Number of items per page
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SummaryActivity'
                title: Response Getathleteactivities
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /activities/{activity_id}/streams:
    get:
      tags:
        - Athlete
      summary: Get Activity Streams
      description: Returns the given activity's streams.
      operationId: getSegmentStreams
      parameters:
        - name: activity_id
          in: path
          required: true
          schema:
            type: integer
            description: The identifier of the activity
            title: Activity Id
          description: The identifier of the activity
        - name: keys
          in: query
          required: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/StreamTypeEnum'
            description: Desired stream types
            title: Keys
          description: Desired stream types
        - name: key_by_type
          in: query
          required: false
          schema:
            type: boolean
            description: Must be true
            default: true
            title: Key By Type
          description: Must be true
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /insights/performance-efficiency/{activity_id}:
    get:
      tags:
        - Insights
      summary: Performance Efficiency
      description: Check if HR efficiency improved within an activity.
      operationId: getPerformanceEfficiency
      parameters:
        - name: activity_id
          in: path
          required: true
          schema:
            type: integer
            description: The identifier of the activity
            title: Activity Id
          description: The identifier of the activity
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Getperformanceefficiency
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /insights/recovery-risk:
    get:
      tags:
        - Insights
      summary: Recovery Risk
      description: Compare 7-day vs 28-day load to detect overtraining.
      operationId: getRecoveryRisk
      parameters:
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Getrecoveryrisk
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    StreamTypeEnum:
      type: string
      enum:
        - time
        - distance
        - latlng
        - altitude
        - velocity_smooth
        - heartrate
        - cadence
        - watts
        - temp
        - moving
        - grade_smooth
      title: StreamTypeEnum
    SummaryActivity:
      properties:
        id:
          type: integer
          title: Id
        resource_state:
          type: integer
          title: Resource State
        external_id:
          anyOf:
            - type: string
          nullable: true
          title: External Id
        upload_id:
          anyOf:
            - type: integer
          nullable: true
          title: Upload Id
        athlete:
          $ref: '#/components/schemas/SummaryAthlete'
        name:
          type: string
          title: Name
        distance:
          type: number
          title: Distance
        moving_time:
          type: integer
          title: Moving Time
        elapsed_time:
          type: integer
          title: Elapsed Time
        total_elevation_gain:
          type: number
          title: Total Elevation Gain
        type:
          type: string
          title: Type
        sport_type:
          type: string
          title: Sport Type
        workout_type:
          anyOf:
            - type: integer
          nullable: true
          title: Workout Type
        start_date:
          type: string
          format: date-time
          title: Start Date
        start_date_local:
          type: string
          format: date-time
          title: Start Date Local
        timezone:
          type: string
          title: Timezone
        utc_offset:
          type: integer
          title: Utc Offset
        start_latlng:
          anyOf:
            - items:
                type: number
              type: array
          nullable: true
          title: Start Latlng
        end_latlng:
          anyOf:
            - items:
                type: number
              type: array
          nullable: true
          title: End Latlng
        location_city:
          anyOf:
            - type: string
          nullable: true
          title: Location City
        location_state:
          anyOf:
            - type: string
          nullable: true
          title: Location State
        location_country:
          anyOf:
            - type: string
          nullable: true
          title: Location Country
        achievement_count:
          type: integer
          title: Achievement Count
        kudos_count:
          type: integer
          title: Kudos Count
        comment_count:
          type: integer
          title: Comment Count
        athlete_count:
          type: integer
          title: Athlete Count
        photo_count:
          type: integer
          title: Photo Count
        map:
          anyOf:
            - additionalProperties: true
              type: object
          nullable: true
          title: Map
        trainer:
          type: boolean
          title: Trainer
        commute:
          type: boolean
          title: Commute
        manual:
          type: boolean
          title: Manual
        private:
          type: boolean
          title: Private
        flagged:
          type: boolean
          title: Flagged
        gear_id:
          anyOf:
            - type: string
          nullable: true
          title: Gear Id
        from_accepted_tag:
          anyOf:
            - type: boolean
          nullable: true
          title: From Accepted Tag
        average_speed:
          type: number
          title: Average Speed
        max_speed:
          type: number
          title: Max Speed
        average_cadence:
          anyOf:
            - type: number
          nullable: true
          title: Average Cadence
        average_watts:
          anyOf:
            - type: number
          nullable: true
          title: Average Watts
        weighted_average_watts:
          anyOf:
            - type: integer
          nullable: true
          title: Weighted Average Watts
        kilojoules:
          anyOf:
            - type: number
          nullable: true
          title: Kilojoules
        device_watts:
          anyOf:
            - type: boolean
          nullable: true
          title: Device Watts
        has_heartrate:
          type: boolean
          title: Has Heartrate
        average_heartrate:
          anyOf:
            - type: number
          nullable: true
          title: Average Heartrate
        max_heartrate:
          anyOf:
            - type: integer
          nullable: true
          title: Max Heartrate
        max_watts:
          anyOf:
            - type: integer
          nullable: true
          title: Max Watts
        pr_count:
          type: integer
          title: Pr Count
        total_photo_count:
          type: integer
          title: Total Photo Count
        has_kudoed:
          type: boolean
          title: Has Kudoed
        suffer_score:
          anyOf:
            - type: integer
          nullable: true
          title: Suffer Score
      type: object
      required:
        - id
        - resource_state
        - athlete
        - name
        - distance
        - moving_time
        - elapsed_time
        - total_elevation_gain
        - type
        - sport_type
        - start_date
        - start_date_local
        - timezone
        - utc_offset
        - achievement_count
        - kudos_count
        - comment_count
        - athlete_count
        - photo_count
        - trainer
        - commute
        - manual
        - private
        - flagged
        - average_speed
        - max_speed
        - has_heartrate
        - pr_count
        - total_photo_count
        - has_kudoed
      title: SummaryActivity
    SummaryAthlete:
      properties:
        id:
          type: integer
          title: Id
        resource_state:
          type: integer
          title: Resource State
        firstname:
          anyOf:
            - type: string
          nullable: true
          title: Firstname
        lastname:
          anyOf:
            - type: string
          nullable: true
          title: Lastname
      type: object
      required:
        - id
        - resource_state
      title: SummaryAthlete
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError